sw64: vdso: fix time calculation
Sunway inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/I56P0Z

--------------------------------

Switch type of wall to monotonic time to u64 so we can get the
correct time.


Acked-by: Xie XiuQi <xiexiuqi@huawei.com>

+++ arch/sw_64/kernel/vdso/vgettimeofday.c
@@ -55,8 +55,8 @@ static __always_inline int do_monotonic_coarse(struct timespec64 *ts,
 		const struct vdso_data *data)
 {
 	u32 start_seq;
-	u32 to_mono_sec;
-	u32 to_mono_nsec;
+	u64 to_mono_sec;
+	u64 to_mono_nsec;
 
 	do {
 		start_seq = vdso_data_read_begin(data);
@@ -124,8 +124,8 @@ static __always_inline int do_monotonic(struct timespec64 *ts,
 {
 	u32 start_seq;
 	u64 ns;
-	u32 to_mono_sec;
-	u32 to_mono_nsec;
+	u64 to_mono_sec;
+	u64 to_mono_nsec;
 
 	do {
 		start_seq = vdso_data_read_begin(data);
